Map generation window, presets, custom inspectors, cleanup
Created preset scriptable object to hold map generation parameters, name, creation time and minimap texture.

Changed map generation class to be a Cave class that holds a preset and will generate the map from it on scene load if it doesn't already exist in the scene.

A map generation editor can be launched from a button on the Cave component.